pyproject.toml
[project]
name = "bigquery-funcs"
version = "0.1.0"
description = "BigQuery functions and/or classes, helper scipts, etc. I commonly use when working with the Python SDK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "functag>=1.2.6",
    "google-cloud-bigquery>=3.29.0",
    "google-cloud-secret-manager>=2.23.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
]
